lint: normalize structural code patterns
Apply reviewed Ruff transformations for redundant branches, collections, conditions, and export ordering.

Refs: #47

+++ tests/correctness/test_bit_ordering.py
@@ -361,7 +361,7 @@ class TestConversionBitOrdering:
                         chi *= 1 - 2 * ((x >> i) & 1)
                 val += coeffs[s] * chi
             # Convert from {-1, +1} to {0, 1}
-            reconstructed.append(int(round((1 - val) / 2)))
+            reconstructed.append(round((1 - val) / 2))
 
         assert (
             reconstructed == tt
